% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{internal.GetAndSetLocalKeys}
\alias{internal.GetAndSetLocalKeys}
\title{Fetches and sets access keys from ~/.aws/credentials. Uses the first
keypair found.}
\usage{
internal.GetAndSetLocalKeys()
}
\description{
Fetches and sets access keys from ~/.aws/credentials. Uses the first
keypair found.
}
\keyword{internal}
