% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{internal.GetURLForRole}
\alias{internal.GetURLForRole}
\title{Gets the instance metadata URL for a given IAM role. This function only
generates the URL, you must query it separately to verify if the role exists.}
\usage{
internal.GetURLForRole(role)
}
\description{
Gets the instance metadata URL for a given IAM role. This function only
generates the URL, you must query it separately to verify if the role exists.
}
\keyword{internal}
