% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{internal.SetAWSEnvironmentVariables}
\alias{internal.SetAWSEnvironmentVariables}
\title{Sets environmental variables for:
  AWS_ACCESS_KEY_ID
  AWS_SECRET_ACCESS_KEY}
\usage{
internal.SetAWSEnvironmentVariables(accessKey, secretKey)
}
\description{
These are standard, used by the aws cli tools and other libraries like cloudyr
}
\keyword{internal}
