% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognito_idp.R
\name{precisely.aws.CognitoIdp.getUsersInGroup}
\alias{precisely.aws.CognitoIdp.getUsersInGroup}
\title{Retrieves a list of the usernames of the users in a given Cognito group.}
\usage{
precisely.aws.CognitoIdp.getUsersInGroup(userPoolId, groupName)
}
\arguments{
\item{userPoolId}{The Cognito User Pool ID.}

\item{groupName}{The name of the Cognito group.}
}
\value{
A character vector of the usernames of the users in the group.
}
\description{
Retrieves a list of the usernames of the users in a given Cognito group.
}
