% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{precisely.aws.IAM.AssumeRole}
\alias{precisely.aws.IAM.AssumeRole}
\title{Fetches and sets access keys for a given IAM role. Role must be authorized
with the EC2 instance. Keys are set as environment variables.}
\usage{
precisely.aws.IAM.AssumeRole(role)
}
\arguments{
\item{role}{The IAM Role to assume.}
}
\description{
Fetches and sets access keys for a given IAM role. Role must be authorized
with the EC2 instance. Keys are set as environment variables.
}
\seealso{
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/iam-roles-for-amazon-ec2.html?shortFooter=true#instance-metadata-security-credentials
}
