% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_utilities.R
\docType{data}
\name{precisely.aws.isEC2}
\alias{precisely.aws.isEC2}
\title{Checks for EC2 Instance Metadata to verify if currently running on EC2}
\format{An object of class \code{logical} of length 1.}
\usage{
precisely.aws.isEC2
}
\value{
Returns TRUE if the metadata service responds, otherwise FALSE
}
\description{
Checks for EC2 Instance Metadata to verify if currently running on EC2
}
\keyword{datasets}
