% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{precisely.aws.S3.delete_object}
\alias{precisely.aws.S3.delete_object}
\title{Delete an S3 object.}
\usage{
precisely.aws.S3.delete_object(object, bucket)
}
\arguments{
\item{object}{The name of the S3 object to delete}

\item{bucket}{The name of the S3 bucket the object is located in.}
}
\value{
A character vector containing the success message.
}
\description{
Delete an S3 object.
}
