% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{precisely.aws.S3.get_object}
\alias{precisely.aws.S3.get_object}
\title{Copy an S3 object to a local file.}
\usage{
precisely.aws.S3.get_object(object, bucket, file)
}
\arguments{
\item{object}{The name of the S3 object to download.}

\item{bucket}{The name of the S3 bucket the object is located in.}

\item{file}{The name of, or full path to, the local file to save the downloaded object in.}
}
\value{
A character vector containing the success message.
}
\description{
Copy an S3 object to a local file.
}
