% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{precisely.aws.S3.list_objects}
\alias{precisely.aws.S3.list_objects}
\title{List objects and common prefixes under a specified S3 bucket.}
\usage{
precisely.aws.S3.list_objects(bucket, recursive = TRUE)
}
\arguments{
\item{bucket}{The name of the S3 bucket.}

\item{recursive}{If TRUE, will recursively list objects in a bucket.
That is, rather than showing PRE dirname/ in the output, all the content in a bucket will be
listed in order. Optional, default is TRUE.}
}
\value{
A \code{\link[tibble]{tibble}} of the objects in the bucket and their associated file information.
}
\description{
List objects and common prefixes under a specified S3 bucket.
}
