% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{precisely.aws.S3.put_object}
\alias{precisely.aws.S3.put_object}
\title{Copy a local file to an S3 object.}
\usage{
precisely.aws.S3.put_object(file, object, bucket)
}
\arguments{
\item{file}{The name of, or full path to, the file to upload.}

\item{object}{The name the object should have in S3 (optional, defaults to the file name).}

\item{bucket}{The name of the S3 bucket to create the object in.}
}
\value{
A character vector containing the success message.
}
\description{
Copy a local file to an S3 object.
}
