% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_manager.R
\name{precisely.aws.SecretsManager.getSecretValue}
\alias{precisely.aws.SecretsManager.getSecretValue}
\title{Retrieves the contents of the encrypted field SecretString from the specified version of a secret
using the AWS CLI.}
\usage{
precisely.aws.SecretsManager.getSecretValue(secretId, versionId,
  versionStage)
}
\arguments{
\item{secretId}{The ARN or the friendly name of the secret.}

\item{versionId}{The ID of of the version of the secret to retrieve (optional).}

\item{versionStage}{The staging label attached to the version of the secret to retrieve (optional).}
}
\value{
list of the parsed JSON of the SecretString
}
\description{
Retrieves the contents of the encrypted field SecretString from the specified version of a secret
using the AWS CLI.
}
\seealso{
https://docs.aws.amazon.com/cli/latest/reference/secretsmanager/get-secret-value.html
}
